Feat : 약 알람 개수 리스트로 dto, entity 변경   (#17)

+++ src/main/java/com/kgu/life_watch/domain/notification/dto/MedicineAlarmDto.java
@@ -13,7 +13,7 @@ public record MedicineAlarmDto(
     String medicineNote,
     boolean completed,
     String repeatCycle,
-    int dosage) {
+    double dosage) {
 
   public static MedicineAlarmDto fromEntity(MedicineAlarm alarm) {
     return new MedicineAlarmDto(
